ddo path
